vhci: extended resume debug logging
includes delays so its possible to take a video of `journalctl -kf`
while apple-bce-vhci resumes, such that you can record the logs, even
though linux might not be able to sync them to disk if there is a crash.

Also made vhci do a suspend/resume cycle when it loads so you can see
said logs without having to figure out how to make the display not
suspend.